удалён лишний файл + добавлена 2 часть работы
