+++ home_work_11/templates/all_candidates.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Кандидаты</title>
+</head>
+<body>
+<h1>Все кандидаты</h1>
+
+{% for candidate in candidates %}
+<p><a href="/candidate/{{candidate.id}}">{{candidate.name}}</a></p>
+{% endfor %}
+
+</body>
+</html>
