+++ home_work_11/templates/candidate_id.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Кандидат</title>
+</head>
+<body>
+<h1>{{candidate.name}}</h1>
+<p>{{candidate.position}}</p>
+<img src="{{candidate.picture}}" width=200/>
+<p>{{candidate.skills}}</p>
+</body>
+</html>
