+++ home_work_11/templates/skill.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>skill</title>
+</head>
+<body>
+<h1>Найдено со скиллом {{len(candidate_skill)}}: 2</h1>
+{{% for candidate in candidates %}}
+<p><a href="/candidate/{{candidate.id}}">{{candidate.name}}</a></p>
+{% endfor %}
+
+</body>
+</html>
